+++ effects/stream/stream.go
@@ -9,8 +9,8 @@ import (
 	"github.com/on-the-ground/effect_ive_go/effects/concurrency"
 	"github.com/on-the-ground/effect_ive_go/effects/internal/helper"
 	effectmodel "github.com/on-the-ground/effect_ive_go/effects/internal/model"
-	"github.com/on-the-ground/effect_ive_go/effects/internal/orderedbuffer"
 	"github.com/on-the-ground/effect_ive_go/effects/log"
+	"github.com/on-the-ground/effect_ive_go/shared/orderedbuffer"
 	"go.uber.org/zap"
 )
